+++ lib/theme/app_images.dart
@@ -2,4 +2,16 @@ class AppImages {
   AppImages._();
 
   static const logo = 'assets/logo.png';
+  
+  static const s1 = 'assets/seasons/season_1.png';
+  static const s2 = 'assets/seasons/season_2.png';  
+  static const s3 = 'assets/seasons/season_3.png';
+  static const s4 = 'assets/seasons/season_4.png';
+  static const s5 = 'assets/seasons/season_5.png';
+
+  static const l1 = 'assets/dimensions/dimension_unknown.png';
+  static const l2 = 'assets/dimensions/dimension_fantasy.png';
+  static const l3 = 'assets/dimensions/dimension_replacement.png';
+  static const l4 = 'assets/dimensions/dimension_c137.png';
+  
 }